requirements.txt
torch
torchvision
einops
wandb
pyyaml
tqdm
